% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\docType{data}
\name{mobile_app}
\alias{mobile_app}
\title{mobile_app}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
mobile_app
}
\description{
mobile_app
}
\keyword{datasets}
